+++ tests/lib/rules/jquery-dollar-sign-reference.js
@@ -234,6 +234,15 @@ ruleTester.run('jquery-dollar-sign-reference', rule, {
 
     {code: 'var foo = {["bar"]: "bar"}', parserOptions: {ecmaVersion: 2015}},
     {code: 'var foo = {["$bar"]: $bar}', parserOptions: {ecmaVersion: 2015}},
+
+    {code: 'async function foo() { const foo = await aPromise; }', parserOptions: {ecmaVersion: 2017}},
+    {code: 'async function foo() { const foo = await $aPromise; }', parserOptions: {ecmaVersion: 2017}},
+    {code: 'async function foo() { const $foo = await aPromise; }', parserOptions: {ecmaVersion: 2017}},
+    {code: 'async function foo() { const $foo = await $aPromise; }', parserOptions: {ecmaVersion: 2017}},
+    {code: 'async function foo() { const {foo} = await $aPromise; }', parserOptions: {ecmaVersion: 2017}},
+    {code: 'async function foo() { const {$foo} = await bar(); }', parserOptions: {ecmaVersion: 2017}},
+    {code: 'async function foo() { $foo = await aPromise; }', parserOptions: {ecmaVersion: 2017}},
+    {code: 'async function foo() { $foo = await bar(); }', parserOptions: {ecmaVersion: 2017}},
   ],
   invalid: [
     {